+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * **\[processor\]** Generate a compile time error if a `@Prop` annotated method returns a boxed primitive and
   is annotated with `@Nonnull` annotation. The code should just use the primitive type.
+* **\[processor\]** Generate a compile error if a component that does not extend `ReactArezComponent` has a field
+  annotated with an arez annotation.
 
 ### [v0.98](https://github.com/react4j/react4j/tree/v0.98) (2018-10-04)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.97...v0.98)